// src/main/resources/angular/src/app/api/model/serviceCondition.ts
/**
 * MAAS API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ServiceCondition {
    serviceConditionId: number;
    version: number;
    active: number;
    client: string;
    publishDate: string;
    url_ca: string;
    url_es: string;
    url_en: string;
    /**
     * Url with the locale of the user, to be used for the apps.
     */
    url?: string;
}